+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 For a full diff see [`0.6.1...main`][0.6.1...main].
 
+### Added
+
+- Added form to enter reading progress directly in the book details modal
+
 ## [`0.6.1`][0.6.1] - 2025-03-30
 
 For a full diff see [`0.6.0...0.6.1`][0.6.0...0.6.1].