+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-For a full diff see [`0.6.1...main`][0.6.1...main].
+For a full diff see [`0.7.0...main`][0.7.0...main].
+
+## [`0.7.0`][0.7.0] - 2025-04-02
+
+For a full diff see [`0.6.1...0.7.0`][0.6.1...0.7.0].
 
 ### Added
 
@@ -96,10 +100,12 @@ For a full diff see [`0.2.2...0.3.0`][0.2.2...0.3.0].
 [0.5.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.5.0
 [0.6.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.6.0
 [0.6.1]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.6.1
+[0.7.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.7.0
 [0.2.2...0.3.0]: https://github.com/weph/obsidian-bookshelf/compare/0.2.2...0.3.0
 [0.3.0...0.3.1]: https://github.com/weph/obsidian-bookshelf/compare/0.3.0...0.3.1
 [0.3.1...0.4.0]: https://github.com/weph/obsidian-bookshelf/compare/0.3.1...0.4.0
 [0.4.0...0.5.0]: https://github.com/weph/obsidian-bookshelf/compare/0.4.0...0.5.0
 [0.5.0...0.6.0]: https://github.com/weph/obsidian-bookshelf/compare/0.5.0...0.6.0
 [0.6.0...0.6.1]: https://github.com/weph/obsidian-bookshelf/compare/0.6.0...0.6.1
-[0.6.1...main]: https://github.com/weph/obsidian-bookshelf/compare/0.6.1...main
+[0.6.1...0.7.0]: https://github.com/weph/obsidian-bookshelf/compare/0.6.1...0.7.0
+[0.7.0...main]: https://github.com/weph/obsidian-bookshelf/compare/0.7.0...main